% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocheck_code.R
\name{autocheck_code}
\alias{autocheck_code}
\title{Create a final part of the javascript code}
\usage{
autocheck_code()
}
\description{
Create a final part of the javascript code
}
\examples{

# ```{r, results='asis', echo=FALSE}
# autocheck_code()
# ```

}
\seealso{
\code{\link{autocheck_question}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
