% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hint.R
\name{check_hint}
\alias{check_hint}
\title{Create hint for 'rmarkdown'}
\usage{
check_hint(
  hint_text,
  hint_title = "Click here to see/close the hint",
  type = c("onclick", "onmouseover", "ondblclick"),
  hint_id = sample(1e+05:1, 1)
)
}
\arguments{
\item{hint_text}{hint paragraph text; can contain markdown}

\item{hint_title}{click paragraph text; can contain markdown}

\item{type}{character that describes behavior of the hint. Possible values are: \code{onclick}, \code{onmouseover}, \code{ondblclick}}

\item{hint_id}{unique identifier of the question}
}
\value{
returns the html tags and javascript code
}
\description{
Create hint for 'rmarkdown'
}
\examples{

check_hint("You can use argument `echo=FALSE`!")

}
\author{
George Moroz <agricolamz@gmail.com>
}
