% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hints.R
\name{check_hints}
\alias{check_hints}
\title{Create list of hints for 'rmarkdown'}
\usage{
check_hints(
  hint_text,
  hint_title = "Click here to see/close the hint",
  hint_title_prefix = "",
  hint_title_suffix = "",
  list_title = "Click here to see/close the list of hints",
  type = c("onclick", "onmouseover", "ondblclick")
)
}
\arguments{
\item{hint_text}{hint paragraph texts; can contain markdown}

\item{hint_title}{hint title texts; can contain markdown}

\item{hint_title_prefix}{string that added before each hint_title}

\item{hint_title_suffix}{string that added after each hint_title}

\item{list_title}{unique identifier for each hint}

\item{type}{character that describes behavior of the hints. Possible values are: \code{onclick}, \code{onmouseover}, \code{ondblclick}}
}
\value{
returns the html tags and javascript code
}
\description{
Create list of hints for 'rmarkdown'
}
\examples{

check_hints(1:4)

}
\author{
George Moroz <agricolamz@gmail.com>
}
