% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_audio.R
\name{insert_audio}
\alias{insert_audio}
\title{Insert multiple images into questions and hints}
\usage{
insert_audio(
  src,
  controls = TRUE,
  autoplay = FALSE,
  loop = FALSE,
  muted = FALSE,
  preload = c("none", "auto", "metadata")
)
}
\arguments{
\item{src}{URL}

\item{controls}{boolian}

\item{autoplay}{boolian}

\item{loop}{boolian}

\item{muted}{boolian}

\item{preload}{character}
}
\value{
returns set of html tags
}
\description{
Insert multiple images into questions and hints
}
\examples{

insert_audio("1.wav")

}
\author{
George Moroz <agricolamz@gmail.com>
}
