% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_images.R
\name{insert_images}
\alias{insert_images}
\title{Insert multiple images into questions and hints}
\usage{
insert_images(
  src,
  alt = "image",
  image_width = "100\%",
  image_height = "100\%"
)
}
\arguments{
\item{src}{image file names}

\item{alt}{alternate texts for an image in case If a browser cannot find an image}

\item{image_width}{image widths. Could be absolute value in pixels or percantage.}

\item{image_height}{image heights. Could be absolute value in pixels or percantage.}
}
\value{
returns the string with html code
}
\description{
Insert multiple images into questions and hints
}
\examples{

# ```{r, results='asis', echo=FALSE}
# insert_images(c("1.png", "2.png"))
# ```

}
\author{
George Moroz <agricolamz@gmail.com>
}
