% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_score.R
\name{insert_score}
\alias{insert_score}
\title{Create an overall sum score for all correct tasks in the document}
\usage{
insert_score()
}
\value{
returns the html tags and javascript code
}
\description{
Create an overall sum score for all correct tasks in the document
}
\author{
George Moroz <agricolamz@gmail.com>
}
