% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocheck_question.R
\name{autocheck_question}
\alias{autocheck_question}
\title{Create check-fields and check-boxes for 'rmarkdown'}
\usage{
autocheck_question(
  question_id,
  answer,
  right = "Correct",
  wrong = "I have a different answer",
  options = NULL
)
}
\arguments{
\item{question_id}{unique identidier of the question}

\item{answer}{correct answer (can be a double or a string)}

\item{right}{form reaction on right answer}

\item{wrong}{form reaction on wrong answer}

\item{options}{vector of values for the selection list type}
}
\description{
Create check-fields and check-boxes for 'rmarkdown'
}
\examples{

# ```{r, results='asis', echo=FALSE}
# autocheck_question(question_id = 1, answer = 5)
# ```

}
\author{
George Moroz <agricolamz@gmail.com>
}
