% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hints.R
\name{check_hints}
\alias{check_hints}
\title{Create check-fields and check-boxes for 'rmarkdown'}
\usage{
check_hints(
  hint_text,
  hint_title = "Click here to see/close the hint",
  hint_title_prefix = "",
  hint_title_suffix = "",
  list_title = "Click here to see/close the list of hints",
  hint_id
)
}
\arguments{
\item{hint_text}{hint paragraph texts; can contain markdown}

\item{hint_title}{hint title texts; can contain markdown}

\item{hint_title_prefix}{string that added to each hint_title to the left side}

\item{hint_title_suffix}{string that added to each hint_title to the right side}

\item{list_title}{unique identidier for each hint}

\item{hint_id}{unique identidier for each hint}
}
\value{
returns the html and javascript code
}
\description{
Create check-fields and check-boxes for 'rmarkdown'
}
\examples{

# ```{r, results='asis', echo=FALSE}
# check_hints(1:4)
# ```

}
\author{
George Moroz <agricolamz@gmail.com>
}
