% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_audio.R
\name{insert_audio}
\alias{insert_audio}
\title{Insert audio}
\usage{
insert_audio(
  src,
  controls = TRUE,
  autoplay = FALSE,
  loop = FALSE,
  muted = FALSE,
  preload = c("none", "auto", "metadata")
)
}
\arguments{
\item{src}{character. It specifies the location (URL) of the audio file.}

\item{controls}{logical. When \code{TRUE}, it specifies that audio controls should be displayed.}

\item{autoplay}{logical. When \code{TRUE}, the audio will automatically start playing as soon as it can do so without stopping.}

\item{loop}{logical. When \code{TRUE}, it specifies that the audio will start over again, every time it is finished.}

\item{muted}{logical. When \code{TRUE}, it specifies that the audio output should be muted.}

\item{preload}{character. It specifies if and how the author thinks that the audio file should be loaded when the page loads. Possible values are \code{none}, \code{auto} and \code{metadata}}
}
\value{
returns set of html tags
}
\description{
Insert audio
}
\examples{

insert_audio("1.wav")

}
\author{
George Moroz <agricolamz@gmail.com>
}
